files/kafkasubscriber/requirements.txt
kafka-python==1.3.1
Jinja2==2.9.5
